Replace the Sandbox with a Library object
This provides a more specific interface to external libraries. Instead
of adding everything to a global prototype a module requires only what
it needs. These dependencies are then initialized and passed into the
module constructor.

This makes testing modules much easier as the entire sandbox object
is removed from the equation.